Added dynamic operator; Small bugfixes within some error operators
Added a "dynamic" operator, that accepts a factory function to be called every time the iterable is iterated upon. The factory function can return an async iterator, or any other AsyncIterableLike, which is then converted to an async iterator.

Removed some unused imports from the splitErrors operator.

Added two new public properties to the subject iterable, `bufferSize` and `queueSize`.