Improvements to the shared, synchronize and replay operators.
Refactored both`the `synchronize` and `shared` operators to return AsyncIterable instead of an AsyncIterator. This means that the objects returned from these operators can be iterated multiple times, and each time they will create a new iterator based on the source iterable that was provided. The various iterators will be linked to each other based on order of creation.

An auxiliary structure `LinkedIterables` was created to support this behavior. It allows creating a generic state that can be shared by linked iterators on the same level (level is defined by the number of iterators created for each iterable, the first iterator in each iterable is level 1, the second created is level 2, and so on).

The `replay` operator now accepts an `AsyncIterable` instead of an `AsyncIterator`, and only creates the iterator when it is first needed. After that, all other iterators use the same underlying parent iterator. If the replay iterable is never used, then no iterator is created.